requirements.txt
click==8.1.3
flake8==5.0.4
flake8-broken-line==0.6.0
flake8-isort==5.0.0
flake8-plugin-utils==1.3.2
flake8-return==1.1.3
Flask==2.2.2
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
pep8-naming==0.13.2
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.21.0
waitress==2.1.2
Werkzeug==2.2.2
